# public/openapi.yaml
openapi: 3.1.0
info:
  title: Work & Vacation Planner API
  version: 1.0.0
  description: |
    REST API for the Work & Vacation Planner — a homelab tool for tracking
    working days, vacation budgets, public holidays, and project schedules.

    Import this file into n8n or any OpenAPI-compatible client.
    Base URL on your homelab: `http://<your-host>:3000`
servers:
  - url: /api
    description: Relative (same origin)
components:
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Invalid date format (YYYY-MM-DD)
...
    WorkingDaySummary:
      type: object
      required: [from, to, total_days, weekdays, public_holidays, vacation_days, working_days]
      properties:
        from:
          type: string
          format: date
          example: "2025-01-01"
        to:
          type: string
          format: date
          example: "2025-12-31"
        total_days:
          type: integer
          example: 365
        weekdays:
          type: integer
          description: Number of weekdays (Mon–Fri) in the range
          example: 261
        public_holidays:
          type: integer
          description: Public holidays that fall on weekdays
          example: 9
        vacation_days:
          type: integer
          description: Vacation days taken (on working weekdays)
          example: 15
        working_days:
          type: integer
          description: Actual billable working days
          example: 237
        working_weekend_days:
          type: integer
          description: Weekend days explicitly marked as working
          example: 0
        non_working_weekday_days:
          type: integer
          description: Weekdays configured as non-working (e.g. permanent day off)
          example: 0
    ProjectWorkingDaySummary:
      type: object
      required: [project_id, project_name, from, to, total_project_days, holidays, vacation_days, working_days]
      properties:
        project_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        project_name:
          type: string
          example: "Client A"
        from:
          type: string
          format: date
          example: "2025-01-01"
        to:
          type: string
          format: date
          example: "2025-12-31"
        total_project_days:
          type: integer
          description: Days matching the project weekday mask
          example: 200
        holidays:
          type: integer
          example: 7
        vacation_days:
          type: integer
          example: 10
        working_days:
          type: integer
          example: 183
    Project:
      type: object
      required: [id, name, colour, weekdays, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        name:
          type: string
          example: "Client A"
        colour:
          type: string
          description: Hex colour string
          example: "#4f86c6"
        startDate:
          type: string
          format: date
          nullable: true
          example: "2025-01-01"
        endDate:
          type: string
          format: date
          nullable: true
          example: "2025-12-31"
        weekdays:
          type: array
          items:
            type: integer
            minimum: 0
            maximum: 6
          description: "Day-of-week numbers active for this project (0=Sun, 1=Mon … 6=Sat)"
          example: [1, 2, 3, 4, 5]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectDayOverride:
      type: object
      required: [projectId, date, type]
      properties:
        projectId:
          type: string
          format: uuid
        date:
          type: string
          format: date
          example: "2025-06-10"
        type:
          type: string
          enum: [include, exclude]
          example: exclude
    MonthlySummary:
      type: object
      required: [month, weekdays, public_holidays, vacation_days, working_days]
      properties:
        month:
          type: string
          example: "2025-01"
        weekdays:
          type: integer
        public_holidays:
          type: integer
        vacation_days:
          type: integer
        working_days:
          type: integer
    YearlySummary:
      type: object
      required: [year, months]
      properties:
        year:
          type: integer
          example: 2025
        months:
          type: array
          items:
            $ref: "#/components/schemas/MonthlySummary"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
  /working-days:
    get:
      summary: Global or per-project working day stats
      description: Returns a WorkingDaySummary for the given date range. If `project` is provided, returns a ProjectWorkingDaySummary scoped to that project's weekday mask and overrides.
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-01-01"
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-12-31"
        - name: project
          in: query
          required: false
          schema:
            type: string
          description: Project UUID — if provided returns per-project stats
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        "200":
          description: Working day statistics
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/WorkingDaySummary"
                  - $ref: "#/components/schemas/ProjectWorkingDaySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /summary:
    get:
      summary: Full-year monthly breakdown
      description: Returns working day stats broken down by month for a given year.
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
          example: 2025
      responses:
        "200":
          description: Yearly summary with monthly breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearlySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /export:
    get:
      summary: CSV export of all working days for a year
      description: Downloads a CSV file listing every working day (weekdays only, excluding regular weekends) with type, holiday name, and active projects.
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
          example: 2025
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
              example: |
                Date,DayOfWeek,Type,Holiday Name,Projects
                2025-01-01,Wednesday,Holiday,New Year's Day,
                2025-01-02,Thursday,Working,,Client A
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"

    get:
      summary: Get one project
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update a project
      description: Partial update — only include fields you want to change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                colour:
                  type: string
                start_date:
                  type: string
                  format: date
                  nullable: true
                end_date:
                  type: string
                  format: date
                  nullable: true
                weekdays:
                  type: array
                  items:
                    type: integer
                    minimum: 0
                    maximum: 6
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      summary: Delete a project
      description: Deletes the project and all its day overrides (cascade).
      responses:
        "204":
          description: Project deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /projects/{id}/overrides:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: List day overrides for a project
      responses:
        "200":
          description: Array of overrides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectDayOverride"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      summary: Upsert a day override
      description: |
        - `include` — force this date into the project (even if weekday mask wouldn't match)
        - `exclude` — force this date out of the project (even if weekday mask would match)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, type]
              properties:
                date:
                  type: string
                  format: date
                  example: "2025-06-10"
                type:
                  type: string
                  enum: [include, exclude]
                  example: exclude
      responses:
        "200":
          description: Override upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDayOverride"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...